Dictionary: refactor to its own class
